io_crypt: menu changes

+++ modules/io_crypt/io_crypt.c
@@ -43,19 +43,16 @@ static uint32_t iodev_table = 0;
 static uint32_t iodev_ctx = 0;
 static uint32_t iodev_ctx_size = 0;
 
-/* no encryption active */
-#define CRYPT_MODE_NONE                 0
-
 /* en-/decrypt on the fly with a symmetric cipher and given password */
-#define CRYPT_MODE_SYMMETRIC            1
+#define CRYPT_MODE_SYMMETRIC            0
 
 /* encrypt on the fly using a random key which gets stored using an asymmetric cipher. decryption on PC only. */
-#define CRYPT_MODE_ASYMMETRIC           2
-#define CRYPT_MODE_ASYMMETRIC_PARANOID  3
+#define CRYPT_MODE_ASYMMETRIC           1
+#define CRYPT_MODE_ASYMMETRIC_PARANOID  2
 
 /* same ciphers, but files are stored unencrypted and get encrypted when camera is idle */
-#define CRYPT_MODE_BACKGROUND_SYM       4
-#define CRYPT_MODE_BACKGROUND_ASYM      5
+#define CRYPT_MODE_BACKGROUND_SYM       3
+#define CRYPT_MODE_BACKGROUND_ASYM      4
 
 
 /* dont redirect any file to fake images */
@@ -68,6 +65,7 @@ static uint32_t iodev_ctx_size = 0;
 #define CRYPT_FAKE_ALL                  3
 
 
+static CONFIG_INT("io_crypt.enabled", iocrypt_enabled, 0);
 static CONFIG_INT("io_crypt.mode", iocrypt_mode, 0);
 static CONFIG_INT("io_crypt.fake", iocrypt_fake, 0);
 static CONFIG_INT("io_crypt.block_size", iocrypt_block_size, 4);
@@ -225,7 +223,7 @@ static uint32_t hook_iodev_OpenFile(void *iodev, char *filename, int32_t flags, 
     
     trace_write(iocrypt_trace_ctx, "iodev_OpenFile('%s', %d) = %d", filename, flags, fd);
     
-    if(fd < COUNT(iocrypt_files))
+    if(fd < COUNT(iocrypt_files) && iocrypt_enabled)
     {
         iocrypt_files[fd].crypt_ctx.priv = NULL;
         iocrypt_files[fd].header_size = 0;
@@ -375,6 +373,11 @@ static uint32_t hook_iodev_OpenFile(void *iodev, char *filename, int32_t flags, 
 
 static uint32_t hook_iodev_ReadFile(uint32_t fd, uint8_t *buf, uint32_t length)
 {
+    if(!iocrypt_enabled)
+    {
+        return orig_iodev->ReadFile(fd, buf, length);
+    }
+    
     uint32_t fd_pos = iodev_GetPosition(fd);
     
     /* when there is some encryption active, handle file offset */
@@ -418,7 +421,7 @@ static uint32_t hook_iodev_WriteFile(uint32_t fd, uint8_t *buf, uint32_t length)
 {
     uint32_t ret = 0;
     
-    if(fd < COUNT(iocrypt_files))
+    if(fd < COUNT(iocrypt_files) && iocrypt_enabled)
     {
         uint8_t *work_ptr = buf;
         uint32_t misalign = ((uint32_t)buf) % 8;
@@ -627,40 +630,34 @@ static MENU_SELECT_FUNC(iocrypt_test_speed)
 
 static MENU_UPDATE_FUNC(iocrypt_update)
 {
+    if(!iocrypt_enabled)
+    {
+        MENU_SET_VALUE("OFF");
+        return;
+    }
+    
     switch(iocrypt_mode)
     {
         case CRYPT_MODE_SYMMETRIC:
             if(!iocrypt_key)
             {
-                MENU_SET_VALUE("Password (missing)");
                 MENU_SET_WARNING(MENU_WARN_NOT_WORKING, "Please set a password first. Files are not encrypted yet.");
             }
-            else
-            {
-                MENU_SET_VALUE("Password");
-            }
+            MENU_SET_VALUE("ON, Password");
             break;
             
         case CRYPT_MODE_ASYMMETRIC:
         case CRYPT_MODE_ASYMMETRIC_PARANOID:
             if(!crypt_rsa_get_pub(&iocrypt_rsa_ctx))
             {
-                MENU_SET_VALUE("RSA (no keys)");
                 MENU_SET_WARNING(MENU_WARN_NOT_WORKING, "Please create keys first. Files are not encrypted yet.");
             }
             else if(crypt_rsa_get_priv(&iocrypt_rsa_ctx))
             {
-                MENU_SET_VALUE("RSA (insecure)");
                 MENU_SET_WARNING(MENU_WARN_ADVICE, "Move IO_CRYPT.KEY to a safe place and DELETE from card.");
             }
-            else
-            {
-                MENU_SET_VALUE("RSA");
-            }
-            break;
             
-        default:
-            MENU_SET_VALUE("Off");
+            MENU_SET_VALUE("ON, RSA");
             break;
     }
 }
@@ -670,13 +667,18 @@ static struct menu_entry iocrypt_menus[] =
     {
         .name = "Encryption",
         .update = &iocrypt_update,
-        .priv = &iocrypt_mode,
-        .max = 2, /* the others are not implemented yet */
-        .icon_type = IT_DICE_OFF,
-        .choices = (const char *[]) {"OFF", "Password", "RSA", "RSA (paranoid)", "Background PW", "Background RSA"},
-        .help = "Select the encryption mode. The higher the level, the less comfort you have.",
+        .priv = &iocrypt_enabled,
+        .max = 1,
         .submenu_width = 710,
         .children = (struct menu_entry[]) {
+            {
+                .name = "Encryption mode",
+                .priv = &iocrypt_mode,
+                .max = 1, /* the others are not implemented yet */
+                .icon_type = IT_DICE,
+                .choices = (const char *[]) {"Password", "RSA", "RSA (paranoid)", "Background PW", "Background RSA"},
+                .help = "Select the encryption mode. The higher the level, the less comfort you have.",
+            },
             /*
             {
                 .name = "Show fake images",
@@ -704,7 +706,6 @@ static struct menu_entry iocrypt_menus[] =
                 .name = "Ask for password on startup",
                 .priv = &iocrypt_ask_pass,
                 .max = 1,
-                .icon_type = IT_BOOL,
                 .help = "When enabled it will ask for the encryption password right after camera powerup.",
             },
             {
@@ -802,7 +803,7 @@ static unsigned int iocrypt_init()
     iocrypt_scratch_sem = create_named_semaphore("iocrypt_scratch", 1);
     
     /* ask for the initial password */
-    if(iocrypt_ask_pass && (iocrypt_mode == CRYPT_MODE_SYMMETRIC || iocrypt_mode == CRYPT_MODE_BACKGROUND_SYM))
+    if(iocrypt_ask_pass && iocrypt_enabled && (iocrypt_mode == CRYPT_MODE_SYMMETRIC || iocrypt_mode == CRYPT_MODE_BACKGROUND_SYM))
     {
         trace_write(iocrypt_trace_ctx, "io_crypt: Asking for password");
         iocrypt_enter_pw();
@@ -842,6 +843,7 @@ MODULE_INFO_START()
 MODULE_INFO_END()
 
 MODULE_CONFIGS_START()
+    MODULE_CONFIG(iocrypt_enabled)
     MODULE_CONFIG(iocrypt_mode)
     MODULE_CONFIG(iocrypt_block_size)
     MODULE_CONFIG(iocrypt_ask_pass)
